requirements.txt
shopifyapi==12.6.0
google-genai==1.14.0
jinja2==3.1.6
python-dotenv==1.1.0
requests==2.32.3
Pillow==11.1.0
